(#184) refactor : 리뷰 조회 쿼리 개선 완료, 인터페이스 분리

+++ server/src/main/java/seb39_40/coffeewithme/common/domain/Pagination.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class Pagination {
 
-    private int totalRecordCount;   // 전체 데이터 수
+    private long totalRecordCount;   // 전체 데이터 수
     private int totalPageCount;     // 전체 페이지 수
     private int startPage;          // 첫 페이지 번호
     private int endPage;            // 끝 페이지 번호
@@ -13,7 +13,7 @@ public class Pagination {
     private boolean existPrevPage;  // 이전 페이지 존재 여부
     private boolean existNextPage;  // 다음 페이지 존재 여부
 
-    public Pagination(int totalRecordCount, int size, int page) {
+    public Pagination(long totalRecordCount, int size, int page) {
         if (totalRecordCount > 0) {
             this.totalRecordCount = totalRecordCount;
             this.calculation(size, page);
@@ -23,7 +23,7 @@ public class Pagination {
     private void calculation(int size, int page) {
 
         // 전체 페이지 수 계산
-        totalPageCount = ((totalRecordCount - 1) / size) + 1;
+        totalPageCount = (int) (((totalRecordCount - 1) / size) + 1);
 
         // 현재 페이지 번호가 전체 페이지 수보다 큰 경우, 현재 페이지 번호에 전체 페이지 수 저장
         if (page > totalPageCount) {
